% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2.R
\name{boxplot_payments}
\alias{boxplot_payments}
\title{Boxplot}
\usage{
boxplot_payments(data, payments)
}
\arguments{
\item{data}{a dataframe}

\item{payments}{a string name for variable y}
}
\value{
a boxplot of payments vs DRG codes
}
\description{
Boxplot
}
\examples{
boxplot_payments(DRG_data, "Average.Total.Payments")
}
