% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate Statistics for Medicare Payments}
\usage{
calculate_statistics(data, statistics)
}
\arguments{
\item{data}{a data frame}

\item{statistic}{a type of statistic to calculate}
}
\value{
a table shows \code{statistic} of average Medicare payments over all of the DRG codes
}
\description{
This function calculates \code{statistic} over all of the DRG codes for average Medicare payments
}
\examples{
calculate_statistics(DRG_data, "mean")
calculate_statistics(DRG_data, "median")
calculate_statistics(DRG_data, "standard deviation")

}
