% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{make_boxplot}
\alias{make_boxplot}
\title{Title
This function produce a boxplot from given dataframe and given type of payment.}
\usage{
make_boxplot(df, payment)
}
\arguments{
\item{df}{A data include hospital-specific charges for the more than 3,000 U.S. hospitals that receive Medicare Inpatient Prospective Payment System (IPPS) payments. See https://data.cms.gov/Medicare-Inpatient/Inpatient-Prospective-Payment-System-IPPS-Provider/97k6-zzx3 for more information.}

\item{payment}{Type of payment to plot on a boxplot. Select either "Average.Medicare.Payments", "Average.Covered.Charges" or "Average.Total.Payments".}
}
\value{
g A boxplot.
}
\description{
Title
This function produce a boxplot from given dataframe and given type of payment.
}
