% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_boxplot.R
\name{payment_boxplot}
\alias{payment_boxplot}
\title{Boxplot of payments by DRG Code}
\usage{
payment_boxplot(data, variable)
}
\arguments{
\item{variable}{}
}
\value{
boxplot of \code{variable} based on the dataset \code{data}
}
\description{
This function creates ggplot boxplot of \code{variable} based on the dataset \code{data}
}
\examples{
payment_boxplot(DRG_data, "Average.Total.Payments")

}
