% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2.R
\name{statistics_summary}
\alias{statistics_summary}
\title{DRG statistics}
\usage{
statistics_summary(data, column1, column2, type)
}
\arguments{
\item{data}{a dataframe}

\item{column1}{sort by variable's names (DRG codes)}

\item{column2}{variable with numerical data}

\item{type}{string to indicate mean, median, or standard deviation}
}
\value{
a data table via knitr::kable
}
\description{
DRG statistics
}
\examples{
statistics_summary(DRG_data, "DRG.Definition", "Average.Medicare.Payments", "stdev")
}
