% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{summary_stat}
\alias{summary_stat}
\title{Summary stat}
\usage{
summary_stat(df, type_of_stat)
}
\arguments{
\item{df}{A data include hospital-specific charges for the more than 3,000 U.S. hospitals that receive Medicare Inpatient Prospective Payment System (IPPS) payments. See https://data.cms.gov/Medicare-Inpatient/Inpatient-Prospective-Payment-System-IPPS-Provider/97k6-zzx3 for more information.}

\item{type_of_stat}{Type of stat. Select from either "mean" for mean, "std" for standard deviation or "median" for median.}
}
\value{
res A value from the given type of statistic.
}
\description{
This function return a basic statistic from the given dataframe.
}
\examples{
df <- DRG_data
m <- summary_stat(df, "mean")

}
