% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate Statistics for Medicare Payments}
\usage{
calculate_statistics(data, statistic)
}
\arguments{
\item{statistic}{}
}
\description{
Calculate Statistics for Medicare Payments
}
